reset tempalte to blank app
