+++ client/app-example/(tabs)/_layout.tsx
