+++ client/app-example/(tabs)/explore.tsx
