+++ client/app-example/(tabs)/index.tsx
